+++ android/app/build.gradle
@@ -31,7 +31,7 @@ if (keystorePropertiesFile.exists()) {
 }
 
 android {
-    namespace "com.mycompany.studentmarket"
+    namespace "com.uricsc305.studentmarket"
     compileSdkVersion 36
     ndkVersion "28.2.13676358"
 
@@ -56,7 +56,7 @@ android {
 
     defaultConfig {
         // TODO: Specify your own unique Application ID (https://developer.android.com/studio/build/application-id.html).
-        applicationId "com.mycompany.studentmarket"
+        applicationId "com.uricsc305.studentmarket"
         minSdkVersion 23
         targetSdkVersion 36
         versionCode flutterVersionCode.toInteger()
