+++ android/app/google-services.json
@@ -23,6 +23,25 @@
           "other_platform_oauth_client": []
         }
       }
+    },
+    {
+      "client_info": {
+        "mobilesdk_app_id": "1:248213702035:android:6238ae55b9105fa3b85466",
+        "android_client_info": {
+          "package_name": "com.uricsc305.studentmarket"
+        }
+      },
+      "oauth_client": [],
+      "api_key": [
+        {
+          "current_key": "AIzaSyBlS83UAXsv7kI_JnvolpUC7ux60f41Idg"
+        }
+      ],
+      "services": {
+        "appinvite_service": {
+          "other_platform_oauth_client": []
+        }
+      }
     }
   ],
   "configuration_version": "1"
