+++ android/app/src/androidTest/java/com/uricsc305/studentmarket/MainActivityTest.java
@@ -1,4 +1,4 @@
-package com.mycompany.studentmarket;
+package com.uricsc305.studentmarket;
 
 import androidx.test.rule.ActivityTestRule;
 import dev.flutter.plugins.integration_test.FlutterTestRunner;
