+++ android/app/src/main/kotlin/com/example/my_project/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.mycompany.studentmarket
+package com.uricsc305.studentmarket
 
 import io.flutter.embedding.android.FlutterActivity
