+++ lib/index.dart
@@ -7,4 +7,4 @@ export '/pages/login/login_widget.dart' show LoginWidget;
 export '/pages/profile/profile_widget.dart' show ProfileWidget;
 export '/pages/search_home/search_home_widget.dart' show SearchHomeWidget;
 export '/pages/chat_screen/chat_screen_widget.dart' show ChatScreenWidget;
-export '/messages_list/messages_list_widget.dart' show MessagesListWidget;
+export '/pages/messages_list/messages_list_widget.dart' show MessagesListWidget;
